requirements.txt
numpy
tensorflow-gpu
imageio